agent/requirements.txt
playwright
supabase
pyobjc
playwright-stealth
python-dotenv
